+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 #### Changes to attributes
 
 - feat(attributes): Add sentry.normalized_db_query ([#194](https://github.com/getsentry/sentry-conventions/pull/194))
+- double-write OpenTelemetry's `resource.deployment.environment.name` to `sentry.environment` ([#196](https://github.com/getsentry/sentry-conventions/pull/196))
 
 ## 0.3.0